merged

+++ src/CroptopPublisher.sol
@@ -242,9 +242,7 @@ contract CroptopPublisher {
             // Create the metadata for the payment to specify the tier IDs that should be minted. We create manually the original metadata, following
             // the specifications from the JBMetadataResolver library.
             mintMetadata = JBMetadataResolver.addToMetadata({
-                originalMetadata: abi.encodePacked(
-                    bytes32(0), bytes32(abi.encodePacked(FEE_PROJECT_ID), 2), additionalPayMetadata
-                    ),
+                originalMetadata: abi.encodePacked(bytes32(0), bytes32(abi.encodePacked(uint32(FEE_PROJECT_ID), uint8(2))), additionalPayMetadata),
                 idToAdd: bytes4(bytes20(metadata.dataHook)),
                 dataToAdd: abi.encode(true, tierIdsToMint)
             });
